+++ NEWS.md
@@ -1,6 +1,8 @@
 # stats19 3.0.0 2023-10
 
-* Major update so the package works with the new csv files
+* Major update so the package works with the new csv files (up to 2022)
+* Deprecation of `read_accidents` in favour of `read_collisions` and using consistent `collision` instead of `accidents`.
+* Other minor improvements
 
 # stats19 2.0.1 2022-11